fix#5045 : resolve NumberFormatException thrown for elements with xsi:nil attribute
Moved if(isXsiNil) check for OMElements with and without childElements, that contain xsi:nil=true as an attribute.

This if(isXsiNil) check was included only for objects of a datamapping schema, which would exclude OAwriting objects with xsi:nil=true attribute.
As this check was not included for primitive elements(String/boolean/integer/number), when a number element with above attribute went to 'writeFieldElement' method
a NumberFormatException was thrown since the element value is empty.