Set tenant domain when access swagger.json of API
A regression issue has been occured due to the fix for https://wso2.org/jira/browse/CIDENTITY-25. The way that getting tenant domain in the isTenantActive menthod in the JDBCTenantManager is changed to retrieve it from the tenant flow.
Therefore, the tenat domain has to be set whatever the flow that calls the isTenantActive method. This fix is to set the tenant domain when accessing the swagger.json of an API
https://wso2.org/jira/browse/ESBJAVA-5069