Check clustering pattern before adding inbond EP
This is to fix wso2/product-ei#532

When the "clusteringPattern" is set to "nonWorkerManager", Inbound endpoint with coordination=false doesn't work. It gives an INFO log with "Inbound EP will not run in manager node. Same will run on worker(s)", when adding the inbound EP. As, the clustering pattern is nonWorkerManager, considering the node as manager is wrong.

Therefore the fix is to check for the clusterin pattern and if it is workerManager, and the node is not a worker node (in other words, the node is a manager node), then prints that INFO log to communicate that the EP will not run on the manager node.