Synapse API: Creating a resource for each method
When generating a Synapse API from swagger definition, previously, for each path we create only one resource and if there to be multiple methods in a single path, all the methods are added to that resource.

Now we generate a resource of each method in a path (A resource for each path, method combination).